+++ packages/monaco/src/browser/contrib/merge-editor/components/merge-actions/merge-actions.module.less
